Backend/mapping_pipeline/requirements.txt
# Core processing dependencies (latest compatible versions)
numpy>=1.26.0
opencv-python>=4.9.0.80
Pillow>=10.4.0
scipy>=1.14.0

# 3D reconstruction - Production Grade
# COLMAP from system packages, Python integration via subprocess for maximum compatibility
# open3d and pycolmap built from source in multi-stage Dockerfile
trimesh>=4.4.0  # Core 3D mesh processing (ARM64 compatible)

# Computer vision and ML
scikit-image>=0.24.0
matplotlib>=3.9.0
exifread>=3.0.0
rawpy>=0.18.0
imageio>=2.35.0

# Cloud and storage
boto3>=1.35.0
google-cloud-storage>=2.18.0
azure-storage-blob>=12.23.0

# Database
psycopg2-binary>=2.9.9
redis>=5.1.0
sqlalchemy>=2.0.35

# API and web framework
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
pydantic>=2.9.0
httpx>=0.27.0

# Task queue and background processing
celery>=5.4.0
flower>=2.0.1
kombu>=5.4.0

# Monitoring and logging
prometheus-client>=0.21.0
sentry-sdk>=2.16.0
structlog>=24.4.0

# Utilities
click>=8.1.7
python-dotenv>=1.0.1
pyyaml>=6.0.2
requests>=2.32.0
tqdm>=4.66.5

# GPS and geospatial processing
geopy>=2.4.1
pyproj>=3.7.0

# System monitoring for reconstruction processor
psutil>=6.1.0

# File integrity and security
cryptography>=43.0.0

# Development and testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
black>=24.8.0
flake8>=7.1.0
mypy>=1.11.0